Change enterDB to use enterSong()
The line needed to enter the song into the sqlite3 database was added to
enterDB.  enterSong() was included in the previous commit but the enterDB
process was never changed to use it.

enterSong() was changed to check if the song already exists in the
database based on the file's hash.