Add selectSong into db and songSql to parse
selectSong pulls the information from the database. That information is
then passed to songSql in parse. songSql goes through the tuple and places
it in a song object. That object is then returned to selectSong which is
returned to which method that calls it, probably the front end or a
method that parses through the database.