Add web page to play songs
Change database schema.  The name of the artist_albums table was changed
to artist_album, it was a type.  A date was added in the songs table.

Add a function to db to check if the database file exists.

Change selectSong to selectPlay in db.  Parsing of sqlite output is
changed to happen in selectPlay.  selectPlay is only used to select
song information for the page to play music.

Change an error in enterSong where the song object was never passed to
appendAlbum

Change syntax errors in appendSong, appendArist, and appendAlbum.

Change an error in music where if an empty song object was created it
would fail because the fullpath function doesn't check that the variables
are strings.

Add gettype function to check the file extension.

Change year in pullInfo so it pulls the string of the year not the
ID3Year class or whatever it is.

Change the pullyear function to check if the year is a list or not.

Remove songSql from music.  It may be added again depending on other
things that happen in future development.

Add play function to web in the Songs class.  It will pull the song
from the database based on the id passed to play. It displays based on
the play.tmpl