Add functions to display all songs on Songs page
Change name in songs table in database.sql to title

Create selectSongs.  It pulls all the song titles, artists and the albums
from the database and creates song objects from them

Change name in songs to title

Change songs in web to use selectSongs in db and set t.songs to the list
from selectSongs.