Create json output of random song for AJAX
create maxSongs method in db. This returns the total number of songs in
the database.

Change selectPlay in db to ensure that the id is a string

Create getrandom method in helper. Returns a random integer between 1
and the given number.

Create getDict method in songs. Returns a dictionary of information that
would be needed to change the play to this song

Create nextsong method in Player class in web. Gets the maximum number of
songs. Gets a random number between one and max.  Gets a song object for
the id. Gets a json object for the song object based on getDict. Returns
json object to the user.